RR 1.10 tuning on Dec 21
