+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autonomous/TestAuto.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.autonomous;
 
 import static com.qualcomm.robotcore.util.ElapsedTime.Resolution.SECONDS;
