+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/processors/TwoCameraCustomProcessorOpMode.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.processors;
 
 import android.util.Size;
